modified on sum.java

+++ sum.java
@@ -1,9 +1,9 @@
 public class Sum{
     public static void main(String[]a){
-    
+
         System.out.println(sum(1,1));
     }
-    public static int sum(int x, int y){
+    public static int sum(double x, int y){
         return  x+y;
     }
 }
